requirements.txt
fastapi[standard]
python-multipart
python-dotenv
pandas
numpydantic
# frameworks
scikit-learn
